backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-socketio==5.12.1
motor==3.7.0
pydantic-settings==2.8.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
langchain==0.3.23
langgraph==0.3.25
langchain-openai==0.3.12
slowapi==0.1.9
pytest==8.3.5
httpx==0.28.1
